Initial Commit
This is the working specification for the dust language (working title).
I feel that I have finally reached a stable, consistent, and adaptable
set of features for the language to support. However, I have reached
this point after much experimentation and changing that I have lost in
the interim. I may be able to recover some of it, but not all.