Document Updates
Updated the syntax of NeatTricks.txt and Euler.txt.
Also updated the language of NeatTricks to better reflect the shown code
and to better explain the reasoning behind the code.
Moved the error code blocks in Constants.txt into Errors.txt as they
were technically syntax testing. Some of the early syntax ideas may lay
around but I'm not sure.