Nil Type
Defined the semantical relationship of Nil and nil with other
types/values. Nil is now a special type within Dust that functions
similar to the Maybe monad in Haskell. The Nil type is the default type
of all Dust variables (and nil the default value). Statically typing a
variable now also provides a guarantee that the given variable has a
usable value (though currently there is a disconnect between typing and
initializing).
Deprecated the proposed "string interpolation" syntax as a language
feature. It is now function specific
Various rewordings/rearrangings