Development System Plan
Settled on a general plan for how to begin implementing dust. The basic
idea is to work on the interpreter as-if it were a compiler (build a
lexer, then a compiler, than a vm (possibly) before combining the stages
into the interpreter). The idea is that by doing so, I will gradually
improve the efficiency of each phase (in ways that can benefit the whole
program) while also finding and ironing out kinks and possible undefined
behaviors at the earliest possible moment (where fixing such problems is
the easiest).
Reorganized the file structure to reflect the new plan