Implicit static typing
Changed the static typing syntax to add the possibility of performing
assignment. The "typing" is performed implicitly based on the value of
the given expression.
Possible extension of the 'self' keyword in relation to functions
Updated various sections to ensure conformity in syntax/execution and
take advantage of new features
Improved the wording of various sections