Left First Evaluation
Forced BinOp to evaluate it's arguments in a strict left->right manner.
This entails adding a swap function to stack and EvalState that swaps
the top two elements on the stack (due to the current ordering of
arguments expected by operator functions).
Added a "clear" command to the repl loop.
Added a basic print function to the register block (though there's no
current way to call it).
Various rewordings/comments in Calculator.h