String Data Type
Implemented the capability to store, recall, and operate on strings. The
current implementation is very basic (uses pointers to a static array in
state.cpp) and will be changed in the future.
Renamed "commonType" to "lub" (though this might be reversed in the
future)
Modified any (or most) instances of DustVal = T to instead use DustVal =
Value<T>::value(T) instead (to accommodate the outside storage of
strings)
Added a "size" command to the repl loop to check the usage of the string
array