Type Creation
Refined the process used to create new types (without using EvalState).
Added a type register to store created types in.
Settled on using single inheritance for the purposes of testing though
current language semantics allow multiple inheritance.
Added in a fairly basic dispatch function that resolves inheritance
relationships to call a given function (and errors when it tries to get
the parent of Object).