lub Formalization
Compartmentalized the process of finding the type that defines a
specific function to the findDef method.
Added global lub(l, r, op) that finds the common type of l and r that
defines op. Doesn't currently account for possible conversion
relationships in the inheritance tree (though the semantics of this is
not defined currently).
Updated the tests to achieve a full spread of the current supported
features.
Added in an initial schema for a possible TypeSystem class.