Converter Functions
Merged the syntax for specifying conversions with the syntax for
specifying type methods. The new syntax for specifying a conversion from
type a to type b is to specify a method "b" in type a's table that
performs the conversion. Setting the appropriate flags is performed
automatically (Though there are no protections against using undeclared
types yet, should raise an error).
Renamed TypeSystem::lub to TypeSystem::com
Added a check for l == r in TypeSystem::lub.
Added more type system tests in testing.
Some notes on modifying TypeSystem::lub/com to allow converters to be
inherited (though the logistics and logic of this is also up for
consideration).