TypeSystem work finalization
Changed immDef to isDefd.
Added a description for a wayto determine whether a converter was
selected by com using the current API.
Added exception point in newType when the type name is not a valid type
identifier (currently only checks if the first letter is capitalized).
Renamed dust::impl::Value to dust::impl::Atom. Added Value struct that
holds an Atom instance and the type of the Atom.