String Storage
Removed the old type system tests from testing.cpp
Added a basic interface for storing and recalling strings. This
interface is largely similar to the old method of storing strings with
the exception of the 'set' method, with allows reassignment of
str_records (and copy-on-write if the only reference).
Added GC (garbage collection) class