Testing Framework Update
Reworked the tests to provide a better coverage of dust features (but
not of possible errors).
Templated the Testing framework on the stream output.
Added possible require_excep for testing that a specific exception is
thrown. Added require_true that checks whether the provided code
evaluates to true.
Modified the testing completion messages produced by the testing
framework.
Defined the run_tests method so that it runs over the tests in
Testing\Tests.txt.