PEGTL Submodule
Changed to using git-submodule for including the PEGTL library (and any future libraries). With this change, PEGTL is also updated to the most recent version (though the code still uses the old).
Added a special "Libraries" folder to hold third-party libraries (as opposed to having a separate folder for each).