Testing Construction error and Block AST
Fixed error in Testing where it would not clear the parse_tree in case
an exception was thrown during parsing. Note: The stack was also not
cleared in the case of  a successful (ie. noexcep) run.
Added a "Block Start" node before evaluation for more accurate parsing
of blocks.
Defined action for block. action<block> takes into account the Block
Start node