Table Work
Solidified work on Continuation Expressions as a language object
Added a section on possibly defining tables as an extension of blocks.
In this formulation, tables are practically stateful blocks of code
(have some extra assignment rules). This would formulation would also do
away with the need for a comma (in a lot of code), though the comma may
be reappropriated as an expression separator.
Added a preliminary implementation of Tables based on the recent
formulation (missing key functionality).