Changed block Rule
Deleted the scp and diff_scp rules (their behavior is incorporated in
the block Rule and Action).
Renamed the old block Rule to prog.
Added a new block Rule that encapsulates recursive parsing of scoped
expressions, with new blocks being created and destroyed with scoping
changes.
Added inline_block rule for use with Control Structure rules to allow
for one-line statements to be in-line (technically allows for the
first-line to be in-line).