Preliminary TryCatch Handling
Tested that try catch statements work as currently devised by getting
a.a.a.a.a.a: 3 to throw an exception.
Added a preliminary syntax description for try-catch (though this will
need a grammar rewrite to fully work). Had to generalize the child nodes
of TryCatch to be ASTNodes instead of Blocks
Reorganized the Action header to group like actions