Fixed Issues with Reference Counting
Fixed Issue that would cause String and Table reference counting to be
either unresponsive or out-of-sync with program execution. This issue
could've caused some values to be collected while they are still in use.
Fixed Issue where variables were not subject to reference counting
updates.
Slight ReadME and ToDo updates