Implemented Better copyInstance
Added a conception of "static" and "member" variables to dust types. Any
variables specified in the expected declaration table are classified as
"members" and are copied over to instances. All definitions outside of
the type declaration are treated as statics and are only accessed in
lookup.
Modified Assign::eval, EvalState::set, and impl::Value to add in the new
semantics.
Current execution also implicitly prevents adding member variables to
types (considered setting a static field). This prevention doesn't
extend to the type itself.